_develop/tools/requirements.txt
mysql-connector-python==8.0.32
protobuf==3.20.3
toml==0.10.2
